Add design for breakout board (initiate hardcore hacking mode)
The first design (Rev A PCB) does not work as intended, because:
- Vertical display is squashed (insufficient gain)
- CH3 does not work. Channel selection truth table was incorrect.

Because the operational parameters of the original Tek hybrid are not
all known, hardcore hacking mode is being initiated. These parameters
must be reverse-enginered. The breakout board added here facilitates
this by allowing a circuit to be dynamically adjusted while connected
to a powered "victim" oscilloscope.

This consists of two PCBs:
 1. The "hybrid" PCB, which brings all the signals to a ribbon cable
 2. The "breakout" PCB, which breaks out the signals from the ribbon
 cable into convenient 0.1" headers.